+++ apparat_launcher/prefs.py
@@ -33,13 +33,13 @@ class PreferenceWindow(wx.Frame):
 
         ## Create the tab windows
         tab1 = UITabGeneral(nb)
-        tab2 = UITabStatistics(nb)
-        tab3 = UITabPluginCommands(nb)
+        tab2 = UITabPluginCommands(nb)
+        tab3 = UITabStatistics(nb)
 
         ## Add the windows to tabs and name them.
         nb.AddPage(tab1, "General ")
-        nb.AddPage(tab2, "Statistics ")
-        nb.AddPage(tab3, "Plugins ")
+        nb.AddPage(tab2, "Plugins ")
+        nb.AddPage(tab3, "Statistics ")
 
         ## Set noteboook in a sizer to create the layout
         sizer = wx.BoxSizer()
@@ -59,7 +59,7 @@ class PreferenceWindow(wx.Frame):
 
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# PREFERENCE-TABS- TABS - # https://pythonspot.com/wxpython-tabs/
+# PREFERENCE-TABS
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 class UITabGeneral(wx.Panel):
 
@@ -70,8 +70,9 @@ class UITabGeneral(wx.Panel):
         wx.Panel.__init__(self, parent)
 
         ## show language
-        cur_ini_value_for_language = ini.read_single_value('Language', 'lang')          # get current value from ini
-        t = wx.StaticText(self, -1, "Language: ", (20, 20))
+        ##
+        cur_ini_value_for_language = ini.read_single_value('Language', 'lang') # get current value from ini
+        txt_language = wx.StaticText(self, -1, "Language: ", (20, 20))
 
         languages = [cur_ini_value_for_language]
         combo_box_style = wx.CB_READONLY
@@ -79,19 +80,13 @@ class UITabGeneral(wx.Panel):
         ui__cb_languages.SetFont(wx.Font(12, wx.MODERN, wx.NORMAL, wx.NORMAL, False, u'Sans'))
         ui__cb_languages.SetValue(languages[0])
 
-        ## Hide UI
-        self.cb_enable_hide_ui = wx.CheckBox(self, -1, 'Hide UI after command execution ', (20, 60))
-        cur_ini_value_for_hide_ui_after_command_execution = ini.read_single_value('General', 'hide_ui_after_command_execution')          # get current value from ini
-        if cur_ini_value_for_hide_ui_after_command_execution == "False":
-            self.cb_enable_hide_ui.SetValue(False)
-        else:
-            self.cb_enable_hide_ui.SetValue(True)
-        wx.EVT_CHECKBOX(self, self.cb_enable_hide_ui.GetId(), self.prefs_general_toggle_hide_ui)
-
+        # line_1
+        self.line1 = wx.StaticLine(self, -1, size=(480, 1), style=wx.LI_HORIZONTAL)
 
         ## icon sizes
-        cur_ini_value_for_iconsize = ini.read_single_value('General', 'icon_size')          # get current value from ini
-        t2 = wx.StaticText(self, -1, "Icon Size: ", (20, 20))
+        ##
+        cur_ini_value_for_iconsize = ini.read_single_value('General', 'icon_size') # get current value from ini
+        txt_iconsize = wx.StaticText(self, -1, "Icon Size: ", (20, 20))
 
         available_icon_sizes = ['128', '256']
         combo_box_style = wx.CB_READONLY
@@ -100,17 +95,61 @@ class UITabGeneral(wx.Panel):
         self.ui__cb_iconsizes.SetValue(cur_ini_value_for_iconsize)
         self.ui__cb_iconsizes.Bind(wx.EVT_TEXT, self.on_change_icon_size) # changing
 
+        # line_2
+        self.line2 = wx.StaticLine(self, -1, size=(480, 1), style=wx.LI_HORIZONTAL)
+
+        ## transparency
+        ##
+        cur_ini_value_for_transparency = ini.read_single_value('General', 'transparency') # get current value from ini
+        txt_transparency = wx.StaticText(self, -1, "UI transparency (default 255, needs restart) : ", (20, 20))
+
+        available_transparency_values = ['255', '250', '240', '230', '220']
+        combo_box_style = wx.CB_READONLY
+        self.ui__cb_transparency = wx.ComboBox(self, wx.ID_ANY, u'', wx.DefaultPosition, wx.Size(100, 30), available_transparency_values, style=combo_box_style)
+        self.ui__cb_transparency.SetFont(wx.Font(12, wx.MODERN, wx.NORMAL, wx.NORMAL, False, u'Sans'))
+        self.ui__cb_transparency.SetValue(cur_ini_value_for_transparency)
+        self.ui__cb_transparency.Bind(wx.EVT_TEXT, self.on_change_transparency) # changing
+
+        # line_3
+        self.line3 = wx.StaticLine(self, -1, size=(480, 1), style=wx.LI_HORIZONTAL)
+
+        ## Hide UI
+        ##
+        self.cb_enable_hide_ui = wx.CheckBox(self, -1, 'Hide UI after command execution ', (20, 60))
+        cur_ini_value_for_hide_ui_after_command_execution = ini.read_single_value('General', 'hide_ui_after_command_execution') # get current value from ini
+        if cur_ini_value_for_hide_ui_after_command_execution == "False":
+            self.cb_enable_hide_ui.SetValue(False)
+        else:
+            self.cb_enable_hide_ui.SetValue(True)
+        wx.EVT_CHECKBOX(self, self.cb_enable_hide_ui.GetId(), self.prefs_general_toggle_hide_ui)
 
         ## Layout
+        ##
         general_sizer = wx.BoxSizer(wx.VERTICAL) # define layout container
         general_sizer.AddSpacer(10)
-        general_sizer.Add(t, 0, wx.ALL, border=10)
+        general_sizer.Add(txt_language, 0, wx.ALL, border=10)
         general_sizer.Add(ui__cb_languages, 0, wx.ALL, border=10)
+
         general_sizer.AddSpacer(30)
-        general_sizer.Add(self.cb_enable_hide_ui, 0, wx.ALL, border=10)
-        general_sizer.AddSpacer(30)
-        general_sizer.Add(t2, 0, wx.ALL, border=10)
+        general_sizer.Add(self.line1, 0, wx.CENTRE)
+        general_sizer.AddSpacer(10)
+
+        general_sizer.Add(txt_iconsize, 0, wx.ALL, border=10)
         general_sizer.Add(self.ui__cb_iconsizes, 0, wx.ALL, border=10)
+
+        general_sizer.AddSpacer(30)
+        general_sizer.Add(self.line2, 0, wx.CENTRE)
+        general_sizer.AddSpacer(10)
+
+        general_sizer.Add(txt_transparency, 0, wx.ALL, border=10)
+        general_sizer.Add(self.ui__cb_transparency, 0, wx.ALL, border=10)
+
+        general_sizer.AddSpacer(30)
+        general_sizer.Add(self.line3, 0, wx.CENTRE)
+        general_sizer.AddSpacer(10)
+
+        general_sizer.Add(self.cb_enable_hide_ui, 0, wx.ALL, border=10)
+
         self.SetSizer(general_sizer)
 
 
@@ -119,7 +158,13 @@ class UITabGeneral(wx.Panel):
         tools.debug_output('on_change_icon_size', 'Preference - General - change icon size: '+str(event), 1)
         tools.debug_output('on_change_icon_size', 'New icon size is set to: '+str(self.ui__cb_iconsizes.GetValue())+'px', 1)
         ini.write_single_value('General', 'icon_size', self.ui__cb_iconsizes.GetValue()) # update preference value
-        
+
+
+    def on_change_transparency(self, event):
+        """Handles the value change of transparency dropbox"""
+        tools.debug_output('on_change_transparency', 'Preference - General - change transparency: '+str(event), 1)
+        tools.debug_output('on_change_transparency', 'New transparency value is set to: '+str(self.ui__cb_transparency.GetValue()), 1)
+        ini.write_single_value('General', 'transparency', self.ui__cb_transparency.GetValue()) # update preference value
 
 
     def prefs_general_toggle_hide_ui(self, event):
@@ -133,7 +178,6 @@ class UITabGeneral(wx.Panel):
             ini.write_single_value('General', 'hide_ui_after_command_execution', "False") # update preference value
 
 
-
 class UITabStatistics(wx.Panel):
 
     """Preference Window - Tab: Statistics - Shows usage stats"""
@@ -143,26 +187,26 @@ class UITabStatistics(wx.Panel):
         wx.Panel.__init__(self, parent)
 
         ## show app start counter
-        cur_ini_value_for_apparat_started = ini.read_single_value('Statistics', 'apparat_started')          # get current value from ini
-        t1 = wx.StaticText(self, -1, "Apparat started:\t\t\t"+cur_ini_value_for_apparat_started, (20, 20))
+        cur_ini_value_for_apparat_started = ini.read_single_value('Statistics', 'apparat_started') # get current value from ini
+        txt_stats__apparat_started = wx.StaticText(self, -1, "Apparat started:\t\t\t"+cur_ini_value_for_apparat_started, (20, 20))
 
         ## show execute counter
-        cur_ini_value_for_command_executed = ini.read_single_value('Statistics', 'command_executed')          # get current value from ini
-        t2 = wx.StaticText(self, -1, "Command executed:\t\t"+cur_ini_value_for_command_executed, (20, 40))
+        cur_ini_value_for_command_executed = ini.read_single_value('Statistics', 'command_executed') # get current value from ini
+        txt_stats__command_executed = wx.StaticText(self, -1, "Command executed:\t\t"+cur_ini_value_for_command_executed, (20, 40))
 
         ## show plugin trigger count
-        cur_ini_value_for_plugin_executed = ini.read_single_value('Statistics', 'plugin_executed')          # get current value from ini
-        t3 = wx.StaticText(self, -1, "Plugins executed:\t\t\t"+cur_ini_value_for_plugin_executed, (20, 60))
+        cur_ini_value_for_plugin_executed = ini.read_single_value('Statistics', 'plugin_executed') # get current value from ini
+        txt_stats__plugin_executed = wx.StaticText(self, -1, "Plugins executed:\t\t\t"+cur_ini_value_for_plugin_executed, (20, 60))
 
+        ## Layout
         statistics_sizer = wx.BoxSizer(wx.VERTICAL) # define layout container
         statistics_sizer.AddSpacer(10)
-        statistics_sizer.Add(t1, 0, wx.ALL, border=10)
+        statistics_sizer.Add(txt_stats__apparat_started, 0, wx.ALL, border=10)
         statistics_sizer.AddSpacer(10)
-        statistics_sizer.Add(t2, 0, wx.ALL, border=10)
+        statistics_sizer.Add(txt_stats__command_executed, 0, wx.ALL, border=10)
         statistics_sizer.AddSpacer(10)
-        statistics_sizer.Add(t3, 0, wx.ALL, border=10)
+        statistics_sizer.Add(txt_stats__plugin_executed, 0, wx.ALL, border=10)
         self.SetSizer(statistics_sizer)
-
 
 
 class UITabPluginCommands(wx.Panel):
@@ -193,7 +237,7 @@ class UITabPluginCommands(wx.Panel):
         cb_enable_plugin_misc = wx.CheckBox(self, -1, 'Misc', (20, 60))
         cb_enable_plugin_misc.SetToolTipString(u'Enable other stuff')
         cb_enable_plugin_misc.SetLabel('misc')
-        cur_ini_value_for_plugin_misc = ini.read_single_value('Plugins', 'enable_plugin_misc')          # get current value from ini
+        cur_ini_value_for_plugin_misc = ini.read_single_value('Plugins', 'enable_plugin_misc') # get current value from ini
         if cur_ini_value_for_plugin_misc == 'True':
             cb_enable_plugin_misc.SetValue(True)
         else:
@@ -209,7 +253,7 @@ class UITabPluginCommands(wx.Panel):
         cb_enable_plugin_nautilus = wx.CheckBox(self, -1, 'Nautilus', (20, 60))
         cb_enable_plugin_nautilus.SetLabel('nautilus')
         cb_enable_plugin_nautilus.SetToolTipString(u'Enables quick access to some nautilus locations/places')
-        cur_ini_value_for_plugin_nautilus = ini.read_single_value('Plugins', 'enable_plugin_nautilus')          # get current value from ini
+        cur_ini_value_for_plugin_nautilus = ini.read_single_value('Plugins', 'enable_plugin_nautilus') # get current value from ini
         if cur_ini_value_for_plugin_nautilus == 'True':
             cb_enable_plugin_nautilus.SetValue(True)
         else:
@@ -225,7 +269,7 @@ class UITabPluginCommands(wx.Panel):
         cb_enable_plugin_screenshot = wx.CheckBox(self, -1, 'Screenshot', (20, 60))
         cb_enable_plugin_screenshot.SetLabel('screenshot')
         cb_enable_plugin_screenshot.SetToolTipString(u'Enables simple screenshot functions')
-        cur_ini_value_for_plugin_screenshot = ini.read_single_value('Plugins', 'enable_plugin_screenshot')          # get current value from ini
+        cur_ini_value_for_plugin_screenshot = ini.read_single_value('Plugins', 'enable_plugin_screenshot') # get current value from ini
         if cur_ini_value_for_plugin_screenshot == 'True':
             cb_enable_plugin_screenshot.SetValue(True)
         else:
@@ -241,7 +285,7 @@ class UITabPluginCommands(wx.Panel):
         cb_enable_plugin_internet_search = wx.CheckBox(self, -1, 'Internet-Search', (20, 60))
         cb_enable_plugin_internet_search.SetLabel('search_internet')
         cb_enable_plugin_internet_search.SetToolTipString(u'Enables search for several popular web-services')
-        cur_ini_value_for_plugin_search_internet = ini.read_single_value('Plugins', 'enable_plugin_search_internet')          # get current value from ini
+        cur_ini_value_for_plugin_search_internet = ini.read_single_value('Plugins', 'enable_plugin_search_internet') # get current value from ini
         if cur_ini_value_for_plugin_search_internet == 'True':
             cb_enable_plugin_internet_search.SetValue(True)
         else:
@@ -257,7 +301,7 @@ class UITabPluginCommands(wx.Panel):
         cb_enable_plugin_local_search = wx.CheckBox(self, -1, 'Local-Search', (20, 60))
         cb_enable_plugin_local_search.SetLabel('search_local')
         cb_enable_plugin_local_search.SetToolTipString(u'Enables search for files and folders in users home directory')
-        cur_ini_value_for_plugin_search_local = ini.read_single_value('Plugins', 'enable_plugin_search_local')          # get current value from ini
+        cur_ini_value_for_plugin_search_local = ini.read_single_value('Plugins', 'enable_plugin_search_local') # get current value from ini
         if cur_ini_value_for_plugin_search_local == 'True':
             cb_enable_plugin_local_search.SetValue(True)
         else:
@@ -268,13 +312,12 @@ class UITabPluginCommands(wx.Panel):
         txt_plugin_local_search.SetForegroundColour('#7f8c8d')
         txt_plugin_local_search.SetFont(font)
 
-
         ## Plugin: Session
         ##
         cb_enable_plugin_session = wx.CheckBox(self, -1, 'Session', (20, 60))
         cb_enable_plugin_session.SetLabel('session')
         cb_enable_plugin_session.SetToolTipString(u'Enables several session commands')
-        cur_ini_value_for_plugin_session = ini.read_single_value('Plugins', 'enable_plugin_session')          # get current value from ini
+        cur_ini_value_for_plugin_session = ini.read_single_value('Plugins', 'enable_plugin_session') # get current value from ini
         if cur_ini_value_for_plugin_session == 'True':
             cb_enable_plugin_session.SetValue(True)
         else:
@@ -289,7 +332,7 @@ class UITabPluginCommands(wx.Panel):
         cb_enable_plugin_shell = wx.CheckBox(self, -1, 'Shell', (20, 60))
         cb_enable_plugin_shell.SetLabel('shell')
         cb_enable_plugin_shell.SetToolTipString(u'Enable executing shell commands')
-        cur_ini_value_for_plugin_shell = ini.read_single_value('Plugins', 'enable_plugin_shell')          # get current value from ini
+        cur_ini_value_for_plugin_shell = ini.read_single_value('Plugins', 'enable_plugin_shell') # get current value from ini
         if cur_ini_value_for_plugin_shell == 'True':
             cb_enable_plugin_shell.SetValue(True)
         else:
@@ -301,10 +344,8 @@ class UITabPluginCommands(wx.Panel):
         txt_plugin_shell.SetFont(font)
 
 
-
         ## Link to plugin commands description
         wxHyperlinkCtrl = wx.HyperlinkCtrl(self, -1, 'Plugin command details', constants.APP_URL+'#plugins')
-
 
 
         ## -----------------------
@@ -316,37 +357,37 @@ class UITabPluginCommands(wx.Panel):
         ## core
         pref_sizer.Add(cb_enable_plugin_core, 0, wx.ALL, border=10)
         pref_sizer.Add(txt_plugin_core, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
+        pref_sizer.AddSpacer(7)
 
         ## misc
         pref_sizer.Add(cb_enable_plugin_misc, 0, wx.ALL, border=10)
         pref_sizer.Add(txt_plugin_misc, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
+        pref_sizer.AddSpacer(7)
 
         ## Nautilus
         pref_sizer.Add(cb_enable_plugin_nautilus, 0, wx.ALL, border=10)
         pref_sizer.Add(txt_plugin_nautilus, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
+        pref_sizer.AddSpacer(7)
 
         ## screenshot
         pref_sizer.Add(cb_enable_plugin_screenshot, 0, wx.ALL, border=10)
         pref_sizer.Add(txt_plugin_screenshot, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
+        pref_sizer.AddSpacer(7)
 
         ## search internet
         pref_sizer.Add(cb_enable_plugin_internet_search, 0, wx.ALL, border=10)
         pref_sizer.Add(txt_plugin_internet_search, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
+        pref_sizer.AddSpacer(7)
 
         ## search local
         pref_sizer.Add(cb_enable_plugin_local_search, 0, wx.ALL, border=10)
         pref_sizer.Add(txt_plugin_local_search, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
+        pref_sizer.AddSpacer(7)
 
         ## Session
         pref_sizer.Add(cb_enable_plugin_session, 0, wx.ALL, border=10)
         pref_sizer.Add(txt_plugin_session, 0, wx.ALL, border=10)
-        pref_sizer.AddSpacer(5)
+        pref_sizer.AddSpacer(7)
 
         ## Shell
         pref_sizer.Add(cb_enable_plugin_shell, 0, wx.ALL, border=10)